fix(spec-5.2): code-review round 2 — heap_lock_tuple P0 + waiter-slot leak + cppcheck
Second review round (user) found a real P0 + a P1 (and a fast-gate Security flag):

- P0: heap_lock_tuple, for a TERMINAL remote lock-only holder, fell through to PG-
  native XactLockTableWait on the remote raw xid — the same Blocker C already fixed in
  heap_update/heap_delete, but via the goto-l3 path. A node must never run native
  wait/hint on a remote raw xid (raw xids alias across instances; native wait can hang).
  Fix: in the require_sleep block set cluster_remote_lockonly_terminal when the C4
  cross-node block resolves a terminal remote lock-only holder; skip the native wait
  switch and set HEAP_XMAX_INVALID (instead of UpdateXmaxHintBits on the remote xid),
  mirroring the heap_update/heap_delete fix.

- P1: cluster_tx_enqueue_wait leaked its waiter slot when CHECK_FOR_INTERRUPTS (query
  cancel / SIGTERM) longjmp'd past txw_slot_clear(). Wrap the loop in PG_TRY/PG_CATCH
  so the slot is always released.

- cluster_smgr.c: InvalidateBackendId >> 16 (shifting a negative -> UB, cppcheck
  shiftNegativeLHS that failed fast-gate Security) -> shift through unsigned; the int8
  backend_hi byte is identical and round-trips to InvalidBackendId.

- Document why the X-transfer no-wire drop ignores its bool return (false == buffer
  not validly resident == no stale copy == safe to grant; the image was already copied).

- New test t/281: cross-node SELECT FOR UPDATE vs FOR UPDATE (heap_lock_tuple), the
  D4 path that had no prior e2e coverage; nightly shard 273-280 -> 273-281.

Spec: spec-5.2-cf-liveread-dataplane-and-tx-row-lock-wait.md (Hardening v1.1 H1.7)

+++ src/backend/cluster/storage/cluster_smgr.c
@@ -747,7 +747,12 @@ cluster_smgr_build_smgr_inval_msg(RelFileLocator rlocator, SharedInvalidationMes
 	 * and reconstruct it in inval.c's SHAREDINVALSMGR_ID apply path.
 	 */
 	out->sm.id = SHAREDINVALSMGR_ID;
-	out->sm.backend_hi = InvalidBackendId >> 16;
+	/* Shift through unsigned: InvalidBackendId is -1 and shifting a negative
+	 * value is UB (cppcheck shiftNegativeLHS).  ((uint32) -1) >> 16 == 0xffff,
+	 * which truncates into the int8 backend_hi as -1 — byte-identical to PG's
+	 * CacheInvalidateSmgr() and round-trips back to InvalidBackendId in the
+	 * SHAREDINVALSMGR_ID apply path. */
+	out->sm.backend_hi = ((uint32)InvalidBackendId) >> 16;
 	out->sm.backend_lo = InvalidBackendId & 0xffff;
 	out->sm.rlocator = rlocator;
 }
